Migrate test_mc_dbk_specphot_kern to test_dbk_specphot_kernels.py. The module test_mc_phot_kernels.py is now empty
